% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_z2GH.R
\name{d_z2GH}
\alias{d_z2GH}
\title{Derivative of \code{\link[=z2gh]{z2gh()}} against \code{z}, on the log-scale}
\usage{
d_z2GH(z, g, h)
}
\arguments{
\item{z}{\link[base]{double} \link[base]{vector}, missingness allowed}

\item{g, h}{\link[base]{double} scalars}
}
\description{
Derivative of \code{\link[=z2gh]{z2gh()}} against \code{z}, on the log-scale
}
\note{
Nomenclature follows the parameters of ?scales::new_transform
}
