% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterValue.R
\name{letterValue}
\alias{letterValue}
\alias{letterValue_g}
\alias{letterValue_h}
\alias{letterValue_gh}
\title{Letter-Value Estimation of Tukey \eqn{g}-&-\eqn{h} Distribution}
\usage{
letterValue_g(
  x,
  probs_g = seq.int(from = 0.01, to = 0.3, by = 0.005),
  g_select = "median",
  ...
)

letterValue_h(x, probs_h = seq.int(from = 0.1, to = 0.3, by = 0.005), ...)

letterValue_gh(
  x,
  probs_g = seq.int(from = 0.01, to = 0.3, by = 0.005),
  probs_h = seq.int(from = 0.1, to = 0.3, by = 0.005),
  g_select = "h_optim",
  ...
)
}
\arguments{
\item{x}{\link[base]{double} \link[base]{vector}, observations \eqn{x},
missingness \emph{\strong{not}} allowed.}

\item{probs_g}{\link[base]{double} \link[base]{vector}, probabilities used for estimating \eqn{g} parameter.
Or, use \code{probs_g = FALSE} to implement the constraint \eqn{g=0}
(i.e., an \eqn{h}-distribution is estimated).}

\item{g_select}{..}

\item{...}{additional parameters, currently not in use}

\item{probs_h}{\link[base]{double} \link[base]{vector}, probabilities used for estimating \eqn{h} parameter.
Or, use \code{probs_h = FALSE} to implement the constraint \eqn{h=0}
(i.e., a \eqn{g}-distribution is estimated).}

\item{true}{(optional) \link[base]{double} \link[base]{vector} of \eqn{(B,g,h)},
for function \code{\link[=letterValue_gh]{letterValue_gh()}} with option \code{g_select = 'demo'}}
}
\value{
Functions \code{\link[=letterValue_g]{letterValue_g()}}, \code{\link[=letterValue_h]{letterValue_h()}} and \code{\link[=letterValue_gh]{letterValue_gh()}}
returns a \link[base]{length}-\code{4L} \link[base]{double} \link[base]{vector} of
estimates \eqn{(\hat{A}, \hat{B}, \hat{g}, \hat{h})}.
}
\description{
Letter-value based estimation (Hoaglin, 1985) of
Tukey \eqn{g}-, \eqn{h}- and \eqn{g}-&-\eqn{h} distribution.
All equation numbers mentioned below refer to Hoaglin (1985).
}
\details{
Unexported function \code{lv_g_()} estimates parameter \eqn{g} using equation (10) for \eqn{g}-distribution
and the equivalent equation (31) for \eqn{g}-&-\eqn{h} distribution.

Unexported function \code{lv_B_()} estimates parameter \eqn{B} for Tukey \eqn{g}-distribution
(i.e., \eqn{g\neq 0}, \eqn{h=0}), using equation (8a) and (8b).

Unexported function \code{lv_Bh_()} estimates parameters \eqn{B} and \eqn{h}
with given \eqn{g\neq 0}, using equation (33),
or with given \eqn{g=0}, using equation (26a), (26b) and (27).

Unexported function \code{letterV_Bh()} estimates parameters \eqn{B} and \eqn{h} for Tukey \eqn{h}-distribution,
i.e., with given \eqn{g=0}, using equation (26a), (26b) and (27).

Function \code{\link[=letterValue]{letterValue()}} plays a similar role as \code{fitdistrplus:::start.arg.default},
thus extends \code{fitdistrplus::fitdist} for estimating Tukey \eqn{g}-&-\eqn{h} distributions.
}
\note{
Parameter \code{probs_g} and \code{probs_h} does not have to be truly unique; i.e., \link[base]{all.equal} elements are allowed.
}
\references{
Hoaglin, D.C. (1985). Summarizing Shape Numerically: The \eqn{g}-and-\eqn{h} Distributions.
\doi{10.1002/9781118150702.ch11}
}
\keyword{internal}
