% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterValue_g.R
\name{lv_g_A}
\alias{lv_g_A}
\title{Estimate \eqn{\hat{A}} from Tukey's \eqn{g}- or \eqn{gh}-distribution}
\usage{
lv_g_A(x, probs_g, A_select = c("optim", "median"), ...)
}
\arguments{
\item{x, probs_g}{see function \code{\link[=letterValue_g]{letterValue_g()}}}

\item{A_select}{\code{'median'}, \code{'optim'} (default)}

\item{...}{additional parameters, currently of no use}
}
\description{
To get various \eqn{\hat{A}} from Tukey's \eqn{g}- or \eqn{gh}-distribution,
by minimizing the standard deviations of \eqn{\hat{g}_p}
}
\keyword{internal}
