% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterValue_g.R
\name{lv_g_g}
\alias{lv_g_g}
\title{Estimate \eqn{\hat{g}} from Tukey's \eqn{g}-distribution}
\usage{
lv_g_g(x, A, probs_g, g_select = c("optim", "median", "demo"), ...)
}
\arguments{
\item{x, probs_g}{see function \code{\link[=letterValue_g]{letterValue_g()}}}

\item{A}{\link[base]{double} scalar, estimate \eqn{\hat{A}}}

\item{g_select}{\code{'median'}, \code{'optim'} (default)}

\item{...}{additional parameters, currently of no use}
}
\description{
To get various \eqn{\hat{g}} from Tukey's \eqn{g}-distribution,
by minimizing the difference between \eqn{\hat{B}_U} and \eqn{\hat{B}_L}.
}
\keyword{internal}
