% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterValue_h.R
\name{lv_h_A}
\alias{lv_h_A}
\title{Estimate \eqn{\hat{A}} from Tukey's \eqn{h}-distribution}
\usage{
lv_h_A(
  x,
  probs_h,
  A_select = c("median", "B.optim", "h.optim"),
  do.plot = FALSE,
  ...
)
}
\arguments{
\item{x, probs_h}{see function \code{\link[=letterValue_h]{letterValue_h()}}}

\item{A_select}{\code{'median'} (default), \code{'B.optim'} or \code{'h.optim'}}

\item{do.plot}{..}

\item{...}{additional parameters, currently of no use}
}
\description{
To get various \eqn{\hat{A}} from Tukey's \eqn{h}-distribution,
by optimizing letter-value based estimates on lower- and upper- half spread.
}
\keyword{internal}
