% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyGH.R
\name{TukeyGH}
\alias{TukeyGH}
\alias{dGH}
\alias{rGH}
\alias{qGH}
\alias{pGH}
\title{Tukey \eqn{g}-&-\eqn{h} Distribution}
\usage{
dGH(x, A = 0, B = 1, g = 0, h = 0, log = FALSE, ...)

rGH(n, A = 0, B = 1, g = 0, h = 0)

qGH(p, A = 0, B = 1, g = 0, h = 0, lower.tail = TRUE, log.p = FALSE)

pGH(q, A = 0, B = 1, g = 0, h = 0, lower.tail = TRUE, log.p = FALSE, ...)
}
\arguments{
\item{x, q}{\link[base]{double} \link[base]{vector}, quantiles}

\item{A}{\link[base]{double} scalar, location parameter \eqn{A}, default \eqn{A=0},}

\item{B}{\link[base]{double} scalar, scale parameter \eqn{B>0}, default \eqn{B=1}}

\item{g}{\link[base]{double} scalar, skewness parameter \eqn{g}, default \eqn{g=0} (i.e., no skewness)}

\item{h}{\link[base]{double} scalar, elongation parameter \eqn{h\geq 0}, default \eqn{h=0} (i.e., no elongation)}

\item{log, log.p}{\link[base]{logical} scalar, if \code{TRUE}, probabilities \eqn{p} are given as \eqn{\log(p)}.}

\item{...}{additional parameters of function \code{\link[=gh2z]{gh2z()}}}

\item{n}{\link[base]{integer} scalar, number of observations}

\item{p}{\link[base]{double} \link[base]{vector}, probabilities}

\item{lower.tail}{\link[base]{logical} scalar, if \code{TRUE} (default), probabilities are \eqn{Pr(X\le x)} otherwise, \eqn{Pr(X>x)}.}
}
\value{
Density function \code{\link[=dGH]{dGH()}} returns \link[base]{double} \link[base]{vector}.

Distribution function \code{\link[=pGH]{pGH()}} returns \link[base]{double} \link[base]{vector}.

Quantile function \code{\link[=qGH]{qGH()}} returns \link[base]{double} \link[base]{vector}.

Random generator function \code{\link[=rGH]{rGH()}} returns \link[base]{double} \link[base]{vector}.
}
\description{
Density, distribution function, quantile function and simulation
for Tukey \eqn{g}-&-\eqn{h} distribution with
location parameter \eqn{A},
scale parameter \eqn{B},
skewness \eqn{g} and
elongation \eqn{h}.
}
\keyword{internal}
