% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{TukeyGH77-package}
\alias{TukeyGH77}
\alias{TukeyGH77-package}
\title{TukeyGH77: Tukey g-&-h Distribution}
\description{
Density, cumulative density, quantile and simulation of the 4-parameter Tukey g-and-h (1977) distributions. The quantile-based transformation (Hoaglin 1985 \doi{10.1002/9781118150702.ch11}) and its reverse transformation.
}
\references{
Tukey, J.W. (1977): Modern Techniques in Data Analysis.
In: NSF-sponsored Regional Research Conference at Southeastern Massachusetts University,
North Dartmouth, MA.

Hoaglin, D.C. (1985): Summarizing shape numerically: The \eqn{g}-and-\eqn{h} distributions.
Exploring data tables, trends, and shapes, pp. 461–513.
John Wiley & Sons, Ltd, New York.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})
}

}
