% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{TukeyGH_helper}
\alias{TukeyGH_helper}
\alias{.dGH}
\alias{.qGH2z}
\title{Helper Functions}
\usage{
.dGH(
  x,
  A,
  B,
  g,
  h,
  log,
  interval = c(-50, 50),
  tol = .Machine$double.eps^0.25,
  maxiter = 1000
)

.qGH2z(
  q,
  q0 = (q - A)/B,
  A = 0,
  B = 1,
  g = 0,
  h = 0,
  interval = c(-15, 15),
  tol = .Machine$double.eps^0.25,
  maxiter = 1000
)
}
\arguments{
\item{x, q}{..}

\item{A, B, g, h}{..}

\item{log}{..}

\item{interval}{..}

\item{tol, maxiter}{..}

\item{q0}{..}
}
\value{
Returns of the helper functions are not documented, for now.
}
\description{
Helper functions to be used in downstream packages.
}
