% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2qGH.R
\name{z2qGH}
\alias{z2qGH}
\title{Tukey \eqn{g}-&-\eqn{h} Transformation}
\usage{
z2qGH(z, A = 0, B = 1, g = 0, h = 0)
}
\arguments{
\item{z}{\link[base]{double} scalar or \link[base]{vector}, standard normal quantiles.}

\item{A, B, g, h}{\link[base]{double} scalar or \link[base]{vector}}
}
\value{
Function \link{z2qGH} returns a \link[base]{double} scalar or \link[base]{vector}.
}
\description{
..
}
\details{
Function \link{z2qGH} transforms standard normal quantiles to Tukey's \eqn{g}-&-\eqn{h} quantiles.
}
