% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterValue_g.R, R/letterValue_gh.R,
%   R/letterValue_h.R
\name{letterValue_g}
\alias{letterValue_g}
\alias{letterValue}
\alias{letterValue_gh}
\alias{letterValue_h}
\title{Letter-Value Estimation of Tukey \eqn{g}-&-\eqn{h} Distribution}
\usage{
letterValue_g(
  x,
  probs_g = seq.int(from = 0.01, to = 0.3, by = 0.005),
  A_select = c("optim", "median", "demo"),
  g_select = c("optim", "median", "demo"),
  ...
)

letterValue_gh(
  x,
  probs_g = seq.int(from = 0.01, to = 0.3, by = 0.005),
  probs_h = seq.int(from = 0.1, to = 0.3, by = 0.005),
  A_select = c("optim", "median", "demo"),
  g_select = c("h.optim", "median", "B.optim", "demo"),
  ...
)

letterValue_h(
  x,
  probs_h = seq.int(from = 0.1, to = 0.3, by = 0.005),
  A_select = c("median", "B.optim", "h.optim", "demo"),
  ...
)
}
\arguments{
\item{x}{\link[base]{double} \link[base]{vector}, observations \eqn{x},
missingness \emph{\strong{not}} allowed.}

\item{probs_g}{\link[base]{double} \link[base]{vector}, probabilities used for estimating \eqn{g} parameter.
Or, use \code{probs_g = FALSE} to implement the constraint \eqn{g=0}
(i.e., an \eqn{h}-distribution is estimated).}

\item{A_select}{..}

\item{g_select}{..}

\item{...}{additional parameters, currently not in use}

\item{probs_h}{\link[base]{double} \link[base]{vector}, probabilities used for estimating \eqn{h} parameter.
Or, use \code{probs_h = FALSE} to implement the constraint \eqn{h=0}
(i.e., a \eqn{g}-distribution is estimated).}
}
\value{
Functions \code{\link[=letterValue_g]{letterValue_g()}}, \code{\link[=letterValue_h]{letterValue_h()}} and \code{\link[=letterValue_gh]{letterValue_gh()}}
returns a \link[base]{length}-\code{4L} \link[base]{double} \link[base]{vector} of
estimates \eqn{(\hat{A}, \hat{B}, \hat{g}, \hat{h})}.
}
\description{
Letter-value based estimation (Hoaglin, 1985) of
Tukey \eqn{g}-, \eqn{h}- and \eqn{g}-&-\eqn{h} distribution.
All equation numbers mentioned below refer to Hoaglin (1985).
}
\details{
Function \code{\link[=letterValue]{letterValue()}} plays a similar role as \code{fitdistrplus:::start.arg.default},
thus extends \code{fitdistrplus::fitdist} for estimating Tukey \eqn{g}-&-\eqn{h} distributions.
}
\note{
Parameter \code{probs_g} and \code{probs_h} does not have to be truly unique; i.e., \link[base]{all.equal} elements are allowed.
}
\references{
Hoaglin, D.C. (1985). Summarizing Shape Numerically: The \eqn{g}-and-\eqn{h} Distributions.
\doi{10.1002/9781118150702.ch11}
}
\keyword{internal}
