% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{tukey_transform}
\alias{tukey_transform}
\alias{z2gh}
\alias{gh2z}
\title{Tukey \eqn{g}-&-\eqn{h} Transformation}
\usage{
z2gh(z, g = 0, h = 0)

gh2z(q, g = 0, h = 0, ...)
}
\arguments{
\item{z}{\link[base]{double} scalar or \link[base]{vector}, standard normal quantiles \eqn{z_q},
missingness is \emph{\strong{not}} allowed}

\item{g, h}{\link[base]{double} scalars}

\item{q}{\link[base]{double} \link[base]{vector}, Tukey \eqn{g}-&-\eqn{h} quantiles \eqn{q},
with location parameter \eqn{A=0} and scale parameter \eqn{B=1},
missingness is \emph{\strong{not}} allowed}

\item{...}{parameters of function \link[rstpm2]{vuniroot}, other than \code{interval}, \code{lower} and \code{upper}}
}
\value{
Functions \code{\link[=z2gh]{z2gh()}} and \code{\link[=gh2z]{gh2z()}} both return
\link[base]{double} scalar, \link[base]{vector} or \link[base]{matrix}.
}
\description{
Function \code{\link[=z2gh]{z2gh()}} (not compute intensive) transforms standard normal quantiles \eqn{z_q} to Tukey \eqn{g}-&-\eqn{h} quantiles.

Function \code{\link[=gh2z]{gh2z()}} (compute intensive!!) transforms Tukey \eqn{g}-&-\eqn{h} quantiles to standard normal quantiles \eqn{z_q}.
}
\details{
Domain of standard normal quantile \eqn{z} to search for function \code{\link[=gh2z]{gh2z()}} is hard coded as
\eqn{(-8.3, 8.3)}, because \code{stopifnot(identical(pnorm(8.3), 1))}.
}
\keyword{internal}
