% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{TukeyGH77-package}
\alias{TukeyGH77}
\alias{TukeyGH77-package}
\title{TukeyGH77: Tukey g-&-h Distribution}
\description{
Density, cumulative density, quantile and simulation of the 4-parameter Tukey g-and-h (1977) distributions. The quantile-based transformation (Hoaglin 1985 \doi{10.1002/9781118150702.ch11}) and its reverse transformation, as well as the letter-value based estimates (Hoaglin 1985).
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with

\code{vignette('intro', package = 'TukeyGH77')}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Other contributors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID}) [contributor]
}

}
