% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{TukeyGH77-package}
\alias{TukeyGH77}
\alias{TukeyGH77-package}
\title{Tukey \eqn{g}-&-\eqn{h} Distribution}
\value{
Returned values of individual functions are documented separately.
}
\description{
Density, cumulative density, quantile and simulation of
the 4-parameter Tukey (1977) \eqn{g}-&-\eqn{h} distributions.
The quantile-based transformation (Hoaglin 1985)
and its reverse transformation,
as well as the letter-value based estimates (Hoaglin 1985),
are also provided.
}
\references{
Tukey, J.W. (1977): Modern Techniques in Data Analysis.
In: NSF-sponsored Regional Research Conference at Southeastern Massachusetts University,
North Dartmouth, MA.

Hoaglin, D.C. (1985): Summarizing shape numerically: The \eqn{g}-and-\eqn{h} distributions.
Exploring data tables, trends, and shapes, pp. 461–513.
John Wiley & Sons, Ltd, New York.
\doi{10.1002/9781118150702.ch11}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

Other contributors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} [contributor]
}

}
